requirements.txt
alembic==1.4.2
blis==0.4.1
catalogue==1.0.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
conda==4.3.16
cryptography==2.9.2
cycler==0.10.0
cymem==2.0.3
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.0/en_core_web_sm-2.3.0.tar.gz
Flask==1.1.2
Flask-Migrate==2.4.0
Flask-SQLAlchemy==2.4.1
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.1
install==1.3.1
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.15.1
kiwisolver==1.2.0
lxml==4.5.1
Mako==1.1.3
MarkupSafe==1.1.1
matplotlib==3.2.2
mccabe==0.6.1
murmurhash==1.0.2
numpy==1.18.5
pandas==1.0.5
Pillow==7.1.2
plac==1.1.3
preshed==3.0.2
psycopg2-binary==2.8.5
py-openaq==1.1.0
pycodestyle==2.6.0
pycosat==0.6.3
pycparser==2.20
pyflakes==2.2.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
pytz==2020.1
requests==2.24.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
scapy==2.4.3
scikit-learn==0.23.1
scipy==1.4.1
seaborn==0.10.1
six==1.15.0
soupsieve==2.0.1
spacy==2.3.0
SQLAlchemy==1.3.17
srsly==1.0.2
thinc==7.4.1
threadpoolctl==2.1.0
tqdm==4.46.1
urllib3==1.25.9
wasabi==0.7.0
Werkzeug==1.0.1
zipp==3.1.0
